chore: update how to use prompt template
