+++ modules/assistant/rag/template.go
@@ -23,11 +23,42 @@
 
 package rag
 
-import "infini.sh/coco/modules/common"
+import (
+	"github.com/tmc/langchaingo/prompts"
+	"infini.sh/coco/modules/common"
+	"infini.sh/framework/core/errors"
+	"regexp"
+)
 
-func GetTemplateArgs(cfg *common.ModelConfig, defaultTemplate string, defaultVars []string) (template string, inputVars []string) {
-	template = defaultTemplate
-	inputVars = defaultVars
+// extractVariables parses a Go template string and returns a slice of unique variable names
+// used in the {{.variable}} syntax.
+func extractVariables(template string) []string {
+	// Regular expression to match {{ .variable }} patterns
+	re := regexp.MustCompile(`{{\s*\.\s*([a-zA-Z0-9_]+)\s*}}`)
+
+	// Find all matches
+	matches := re.FindAllStringSubmatch(template, -1)
+
+	// Use a map to store unique variable names
+	varsMap := make(map[string]struct{})
+	for _, match := range matches {
+		if len(match) > 1 {
+			varsMap[match[1]] = struct{}{}
+		}
+	}
+
+	// Convert map keys to a slice
+	vars := make([]string, 0, len(varsMap))
+	for v := range varsMap {
+		vars = append(vars, v)
+	}
+
+	return vars
+}
+
+func GetPromptByTemplateArgs(cfg *common.ModelConfig, defaultTemplate string, requiredVars []string, inputValues map[string]any) (*prompts.PromptTemplate, error) {
+	template := defaultTemplate
+	inputVars := requiredVars
 
 	if cfg.PromptConfig != nil {
 		if cfg.PromptConfig.PromptTemplate != "" {
@@ -38,5 +69,26 @@ func GetTemplateArgs(cfg *common.ModelConfig, defaultTemplate string, defaultVar
 			inputVars = cfg.PromptConfig.InputVars
 		}
 	}
-	return template, inputVars
+
+	variables := extractVariables(template)
+	missingVars := map[string]interface{}{}
+	for _, v := range variables {
+		if _, exists := inputValues[v]; !exists {
+			missingVars[v] = ""
+		}
+	}
+
+	if len(missingVars) > 0 && len(requiredVars) > 0 {
+		for _, v := range requiredVars {
+			_, ok := missingVars[v]
+			if ok {
+				return nil, errors.Errorf("var [%v] required, but was not found", v)
+			}
+		}
+	}
+
+	prompt := prompts.NewPromptTemplate(template, inputVars)
+	prompt.PartialVariables = missingVars //default value for missing variable
+
+	return &prompt, nil
 }
